fix: dto에 필드 타입 long으로 수정

+++ src/main/java/com/example/wini/domain/log/dto/response/SimpleActionChange.java
@@ -1,6 +1,6 @@
 package com.example.wini.domain.log.dto.response;
 
-public record SimpleActionChange(String text, int monthlyChange) {
+public record SimpleActionChange(String text, long monthlyChange) {
     public static SimpleActionChange from(ActionChange actionChange) {
         return new SimpleActionChange(actionChange.action().getText(), actionChange.monthlyChange());
     }
